restructure forminput component
